+++ webDemo/src/main/java/com/dxc/application/feature/gimmaster/dto/GimHeaderResultDTO.java
@@ -9,7 +9,7 @@ import java.util.Date;
 public class GimHeaderResultDTO {
     private String gimType;
     private String gimDesc;
-    private BigDecimal cdLength;
+    private Integer cdLength;
     private String field1Label;
     private String field2Label;
     private String field3Label;
